stability/requirements.txt
uvicorn
fastapi
pydantic
ouro-py
python-dotenv
requests
pillow